Move `@ts-expect-error` to more precise code locations
These error expectations aren't in precise spots which is causing them to be silenced by unrelated errors. For example, there is nothing currently that checks if the actionHandler state is in the state map and there is nothing that checks the action is in the action map, but these tests are passing because the stateName isn't declared as const and that results in an unrelated error. Let's expose the problems in these tests by making them fail in this commit.